requirements.txt
geopy==2.2.0
exifread==2.1.2
pillow==10.2.0